requirements.txt
feedparser==6.0.11
httpx==0.27.2
groq==0.11.0
jinja2==3.1.4
python-dotenv==1.0.1
